refactor: useCoupon 및 useProduct 훅에서 경로 수정
