+++ src/refactoring/hooks/index.ts
@@ -1,4 +1,2 @@
 export * from './useCart.ts';
-export * from './useCoupon.ts';
-export * from './useProduct.ts';
 export * from './useLocalStorage.ts';
